PrintLastDigit

+++ functions_nested_loops/7-print_last_digit.c
@@ -7,7 +7,7 @@
 int print_last_digit(int n)
 {
 
-	long x = n;
+	int x = n;
 
 	if (x < 0)
 		x = -x;
